pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-marshmallow>=1.3.0",
    "flask-migrate>=4.1.0",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "marshmallow-sqlalchemy>=1.4.2",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
